Added bank bond demand strategy where demand depends on banks deposits
With this strategy a bank's bond demand is determined by their deposits
in relation to total deposits. Additionally, if the money supply of non
banks contains not only deposits, only the portion of deposits to the
non bank money supply of the bonds is distributed via the ratio of
deposits to total deposits. The rest of the bonds is distributed equally
among all banks.